Update to Teads SDK 4.0.12

+++ TeadsSDKDemo/app/src/main/java/tv/teads/teadssdkdemo/format/inread/InReadScrollViewFragment.java
@@ -73,7 +73,7 @@ public class InReadScrollViewFragment extends BaseFragment implements TeadsListe
     }
 
     @Override
-    public void displayAd(TeadsAd teadsAd, int i) {
+    public void displayAd(TeadsAd teadsAd, float i) {
 
     }
